requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
defusedxml==0.7.1
fastapi==0.116.2
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
httpie==3.2.4
idna==3.10
Jinja2==3.1.6
llvmlite==0.45.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
numba==0.62.0
numpy==2.3.3
openai-whisper==20250625
pydantic==2.11.9
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PySocks==1.7.1
python-ffmpeg==2.0.12
python-multipart==0.0.20
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
setuptools==80.9.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tiktoken==0.11.0
torch==2.8.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
